Convert application folder
